Refactor: Exception과 BaseBallGame으로 클래스 분리
